+++ actix-http/CHANGES.md
@@ -6,7 +6,12 @@
 
 * Add support for sending HTTP requests with `Rc<RequestHead>` in addition to sending HTTP requests with `RequestHead`
 
+* Allow to use `std::convert::Infallible` as `actix_http::error::Error`
+
+
 ### Fixed
+
+* h2 will use error response #1080
 
 * on_connect result isn't added to request extensions for http2 requests #1009